+++ src/settings/mod.rs
@@ -1,3 +1,4 @@
+pub mod change_keybinding;
 pub mod creds;
 pub mod keybindings;
 pub mod theme;