+++ CHANGELOG.md
@@ -1,20 +1,34 @@
 # Changelog
 
+## [0.2.2] - 2024-07-06
+
+### Fixed
+
+- Made a seperate struct for keybindings
+
+### Added
+
+- User can now change keybinding through the terminal itself
+
 ## [0.2.1] - 2024-06-08
 
 ### Fixed
+
 - Add to Playlist menu will not dissapear after adding track to playlist
 - Seperated key events in different files
 
 ## [0.2.0] - 2024-06-07
 
 ### Fixed
+
 - Search results would not refresh the second time
 
 ### Added
+
 - User can now add tracks to their playlists.
 - User can now follow/unfollow playlists.
 
 # What is this?
 
 All notable changes to this project will be documented in this file.
+